requirements.txt
fastapi
uvicorn
python-multipart
langchain
chromadb
sentence-transformers
pypdf
groq
python-dotenv
